# docs/swagger.yaml
definitions:
...
  model.CreateTeacherInput:
    properties:
      employee_id:
        type: integer
    type: object
...
  model.Teacher:
    properties:
      address:
        type: string
      birth_date:
        type: string
      email:
        type: string
      employee_id:
        type: integer
      hire_date:
        type: string
      last_visit_at:
        type: string
      name:
        type: string
      notes:
        type: string
      password:
        type: string
      person_id:
        type: integer
      phone:
        type: string
      photo:
        items:
          type: integer
        type: array
      registered_at:
        type: string
      salary:
        type: number
      surname:
        type: string
      teacher_id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  model.UpdateTeacherInput:
    properties:
      employee_id:
        type: integer
    type: object
  model.UserSignInInput:
    properties:
      login:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - login
    - password
    type: object
  model.UserSignUpInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 70
        minLength: 5
        type: string
    required:
    - email
    - password
    - username
    type: object
  service.Tokens:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for test.
  title: University System API
  version: "2.0"
paths:
...
  /admin/hub/teachers:
    get:
      consumes:
      - application/json
      description: get teachers
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            items:
              $ref: '#/definitions/model.Teacher'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get teachers
      tags:
      - admin-teachers
    post:
      consumes:
      - application/json
      description: create teacher
      parameters:
      - description: Data for creating teacher
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/model.CreateTeacherInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create teacher
      tags:
      - admin-teachers
  /admin/hub/teachers/{id}:
    delete:
      consumes:
      - application/json
      description: delete teacher
      parameters:
      - description: ID for deleting teacher
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete teacher
      tags:
      - admin-teachers
    get:
      consumes:
      - application/json
      description: get teacher by id
      parameters:
      - description: ID for getting teacher
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            $ref: '#/definitions/model.Teacher'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get teacher
      tags:
      - admin-teachers
    patch:
      consumes:
      - application/json
      description: update teacher
      parameters:
      - description: ID for updating teacher
        in: path
        name: id
        required: true
        type: integer
      - description: New information for update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateTeacherInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update teacher
      tags:
      - admin-teachers
  /refresh:
    get:
      description: Refresh access token using refresh token stored in cookie
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/service.Tokens'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Refresh Access Token
      tags:
      - auth
  /sign-in:
    post:
      consumes:
      - application/json
      description: Sign in a user and generate access token
      parameters:
      - description: Data for signing in user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserSignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/service.Tokens'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Sign In User
      tags:
      - auth
  /sign-up:
    post:
      consumes:
      - application/json
      description: User registration
      parameters:
      - description: Data for registration user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserSignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Registered
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: User registration
      tags:
      - auth
swagger: "2.0"